+++ CS416FinalProject/settings.py
@@ -140,3 +140,12 @@ CRISPY_TEMPLATE_PACK = 'bootstrap4'
 LOGIN_REDIRECT_URL = '/'
 
 LOGOUT_REDIRECT_URL = '/account/login/'
+
+# Django Resized (we use this for profile pictures)
+DJANGORESIZED_DEFAULT_SIZE = [256, 256]
+DJANGORESIZED_DEFAULT_SCALE = 1.0
+DJANGORESIZED_DEFAULT_QUALITY = 100
+DJANGORESIZED_DEFAULT_KEEP_META = False
+DJANGORESIZED_DEFAULT_FORCE_FORMAT = 'PNG'
+DJANGORESIZED_DEFAULT_FORMAT_EXTENSIONS = {'PNG': ".png"}
+DJANGORESIZED_DEFAULT_NORMALIZE_ROTATION = False
