+++ accounts/forms.py
@@ -1,5 +1,5 @@
 from django import forms
-from .models import User
+from django.contrib.auth.models import User
 
 class LoginForm(forms.ModelForm):
