+++ cosmos/admin.py
@@ -1,8 +1,8 @@
 from django.contrib import admin
-from .models import Post, Follow, Reaction, Profile
+from .models import Post, Follow, Reaction
+from accounts.models import Profile
 
 # Register your models here.
 admin.site.register(Post)
 admin.site.register(Follow)
 admin.site.register(Reaction)
-admin.site.register(Profile)
