+++ cosmos/urls.py
@@ -10,9 +10,6 @@ urlpatterns = [
     path('post/<int:view_post>', views.showPost, name='queryPost'),
     path('user/<int:other_user>', views.showProfile, name='queryUser'),
     path('user/', views.showProfile, name='queryUser'),
-    path('account/edit', views.editUser, name='editUser'),
-    path('account/signup', views.createUser, name='createUser'),
-    path('account/', include('django.contrib.auth.urls'), name="loginPage"),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_URL)
 
 if(settings.DEBUG):
