+++ templates/cosmos/post.html
@@ -45,13 +45,13 @@
   <div class="row d-flex mt-4 justify-content-between">
     <!-- Like -->
     <div class="col d-flex justify-content-center p-0">
-      <a class="btn btn-outline-success" href="" style="width:48px;height:48px;font-size:1.5em;">
+      <a class="btn btn-outline-like" href="" style="width:48px;height:48px;font-size:1.5em;">
         <i class="bi bi-hand-thumbs-up-fill"></i>
       </a>
     </div>
     <!-- Dislike -->
     <div class="col d-flex justify-content-center p-0">
-      <a class="btn btn-outline-danger" href="" style="width:48px;height:48px;font-size:1.5em;">
+      <a class="btn btn-outline-dislike" href="" style="width:48px;height:48px;font-size:1.5em;">
         <i class="bi bi-hand-thumbs-down-fill"></i>
       </a>
     </div>
